Safari-611.1.10/LayoutTests/webgpu/uncaptured-errors-expected.txt

CONSOLE MESSAGE: GPUDevice.createBuffer(): Buffer cannot have both MAP_READ and MAP_WRITE usage!
CONSOLE MESSAGE: GPUDevice.createBuffer(): Buffer cannot have both MAP_READ and MAP_WRITE usage!

PASS Fire an uncaptured GPUDevice error event at 'device'.
PASS Intercept an uncaptured error with GPUDevice.onuncapturederror EventHandler.
PASS GPUUncapturedErrorEvent async test.
